// src/types/firebase/Appreciation.ts
export type Appreciation = {
  id: string
  category: string
  comment_count: number
  created_date: any
  description: string
  due_date: any
  formatted_address: string
  gps_location: any
  image: string
  priority: string
  resident_avatar: string
  resident_id: string
  resident_name: string
  responder_avatar: string
  responder_id: string
  responder_name: string
  status: string
}